JavaLocalVariablesExample
In this project, we will discuss the concept of local variables used in code blocks and methods in Java. Local variables are variables that are only valid within the code block or method in which they are defined. In this example, we will show how local variables are defined and used in a method and how their scope is limited.